Renamed knexHelper to datasourceUtils + removed old code
